requirements-backup.txt
Django==4.2.23
mysqlclient==2.2.6
dj-database-url==2.1.0

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

django-environ==0.11.2
django-cors-headers==4.4.0
django-filter==23.5

drf-spectacular==0.28.0

gunicorn==23.0.0

celery==5.5.3
redis==6.1.1
sentry-sdk==1.45.1

# Storage/CDN and file handling
django-storages==1.14.6
boto3==1.37.38
Pillow==10.4.0

# Notifications / Payments / PDFs
twilio==9.7.0
stripe==12.4.0
# weasyprint==61.2  # Disabled due to system dependency issues

# Timezone utilities
pytz==2025.2

django-redis==5.4.0
